+++ src/scraping5.py
@@ -1,12 +1,50 @@
 from imdb import IMDb
 import pandas as pd
 
-df = pd.read_csv('..\dataset\\netflix_list.csv', sep=",", low_memory=True)
-
 # creating instance of IMDb
 imdb = IMDb()
 
+########################## Trim Title Spaces ##########################
+
+# df = pd.read_csv('..\dataset\\netflix_list.csv', sep=",", low_memory=True)
+
+# df['title'] = df['title'].str.strip()
+# df.to_csv("TitleSpaceTrim.csv", index=False)
+
+############################ Remove Types #############################
+
+# df = pd.read_csv('TitleSpaceTrim.csv', sep=",", low_memory=True)
+
+# # Remove tvEpisodes
+# df = df[df.type != "tvEpisode"]
+
+# # Remove videogame
+# df = df[df.type != "videoGame"]
+
+# # Remove blank types
+# df = df[df.type.notnull()]
+
+# df.to_csv("RemoveTypes.csv", index=False)
+
+############################ Rename Types #############################
+
+# df = pd.read_csv('RemoveTypes.csv', sep=",", low_memory=True)
+
+# df['type'] = df['type'].replace(['tvSeries','tvMiniSeries','tvShort','tvSpecial','tvMovie'],['series','miniSeries','short','special','movie'])
+
+# df.to_csv("RenameTypes.csv", index=False)
+
+####################### Delete Plot and isAdult #######################
+
+# df = pd.read_csv('RenameTypes.csv', sep=",", low_memory=True)
+
+# df = df.drop(columns=['isAdult', 'plot'])
+
+# df.to_csv("DeleteColumns.csv", index=False)
+
 ############################ Scraping Cast ############################
+
+# df = pd.read_csv('DeleteColumns.csv', sep=",", low_memory=True)
 
 # cast_nulls = df.loc[(df['cast'] == "-")]
 # arr = cast_nulls["imdb_id"].to_numpy()
@@ -29,22 +67,12 @@ imdb = IMDb()
 # df.loc[df['cast'] == "-", 'cast'] = arr
 # print(df.loc[df['cast'] == "-"])
 
-
 # # Saves to file
-# df.to_csv("cast.csv", index=False)
+# df.to_csv("ScrapingCast.csv", index=False)
 
-#######################################################################
+########################### Scraping Years ############################
 
-# Removes all TVEpisodes and blank ones
-df = df[df.type != "tvEpisode"]
-df = df[df.type != "videoGame"]
-df = df[df.type.notnull()]
-
-# df.to_csv("RemoveEpisodes.csv", index=False)
-
-############################ Scraping Cast ############################
-
-df = pd.read_csv('..\src\\RemoveEpisodes.csv', sep=",", low_memory=True)
+df = pd.read_csv('ScrapingCast.csv', sep=",", low_memory=True)
 
 no_years = df.loc[(df['startYear'].isna() & df['endYear'].isna()) | (((df['type'] == "series") | (df['type'] == "miniSeries")) & (df['endYear'].isna()))]
 
@@ -52,7 +80,7 @@ def get_years(imdb_id):
     print(imdb_id)
     try:
         movie_serie = imdb.get_movie(imdb_id[2:])
-        movie_series_type = no_years.loc[(no_years['imdb_id'] == imdb_id)] # Ir buscar o type em string
+        movie_series_type = no_years.loc[(no_years['imdb_id'] == imdb_id)].iloc[0]['type'] # Ir buscar o type em string
 
         print(movie_series_type)
         
@@ -71,14 +99,65 @@ print("# Series")
 print(get_years("tt4052886")) # series
 print("# Movie")
 print(get_years("tt1488589")) # movie
-print("# miniSerie")
-print(get_years("tt10048342")) # miniSeries
+print("# MiniSerie")
+print(no_years.loc[(no_years['imdb_id'] == 10048342)])
+print(get_years("tt12837400")) # miniSeries
 print("# Short")
-print(get_years("tt3472226")) # short
+nova = df.loc[(df['startYear'].isna() & df['endYear'].isna() & (df['type'] == 'short')) | (df['type'] == 'series')]
+print(nova.loc[df['type'] == 'short'])
+print(df.loc[(df['startYear'].isna() & df['endYear'].isna() & (df['type'] == 'short')) | (df['type'] == 'series')])
+print(no_years.loc[(no_years['imdb_id'] == 13871260)]) # TODO: Devia entrar no filtro e não entra
+print(get_years("tt13871260")) # short
 print("# Special")
 print(get_years("tt13567480")) # special
 print("# Video")
 print(get_years("tt1107365")) # video
+
+# Saves to file
+# df.to_csv("ScrapingYears.csv", index=False)
+
+########################### Scraping Episodes ############################
+
+# df = pd.read_csv('ScrapingYears.csv', sep=",", low_memory=True)
+
+# # TODO - BeautifulSoup
+
+# # Saves to file
+# df.to_csv("ScrapingEpisodes.csv", index=False)
+
+########################### Scraping Runtimes ############################
+
+# df = pd.read_csv('DeleteColumns.csv', sep=",", low_memory=True) # TODO: Mudar para ScrapingEpisodes
+
+# no_runtime = df.loc[df['runtime']=="\\N"]
+
+# def get_runtimes(imdb_id):
+#     print(imdb_id)
+#     try:
+#         movie_serie = imdb.get_movie(imdb_id[2:])
+#         runtimes = movie_serie['runtimes']
+#         return runtimes[0]
+#     except Exception as e:  
+#         print("Exception getting runtime info from " + imdb_id + ": " + str(e))
+#         return "Not available"
+
+# no_runtime['runtime'] = no_runtime.apply(lambda x: get_runtimes(x['imdb_id']), axis=1)
+# arrRuntimes = no_runtime['runtime'].to_numpy()
+# df.loc[df['runtime'] == "\\N", 'runtime'] = arrRuntimes
+
+# print(arrRuntimes)
+# print("-------------------------------------")
+# print(df.loc[df['runtime'] == "\\N", 'runtime'])
+
+# Saves to file
+# df.to_csv("ScrapingRuntimes.csv", index=False)
+
+# print(get_runtimes("tt2403776"))
+
+# TODO: Talvez mudar de string para int
+
+
+
 
 # localized title
 # original title
@@ -106,8 +185,7 @@ print(get_years("tt1107365")) # video
 # series years
 # akas
 # seasons
-# writer
-# production companies
+# writer          
 # distributors
 # special effects
 # other companies