+++ src/changes.md
@@ -2,5 +2,7 @@
 
 - Removed empty types (2 rows)
 - Removed videoGame type (1 row)
+- Removed tvEpisodes type (785 rows)
 - Joined tvShort and short types
 - Joined tvMovie and movie types
+- Renamed types